- delete some files - finish login-feature
