--- lib/features/add_user_feature/logic/bloc/createuser_event.dart
@@ -1,8 +1,0 @@
-
-import 'package:freezed_annotation/freezed_annotation.dart';
-part 'createuser_event.freezed.dart';
-
-@freezed
-class CreateuserEvent with _$CreateuserEvent {
-  const factory CreateuserEvent.craete() = Create;
-}
