--- lib/features/add_user_feature/logic/bloc/createuser_state.freezed.dart
@@ -1,575 +1,0 @@
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
-
-part of 'createuser_state.dart';
-
-// **************************************************************************
-// FreezedGenerator
-// **************************************************************************
-
-T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
-
-/// @nodoc
-mixin _$CreateuserState {
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() loading,
-    required TResult Function() success,
-    required TResult Function(String message) failure,
-  }) =>
-      throw _privateConstructorUsedError;
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? loading,
-    TResult? Function()? success,
-    TResult? Function(String message)? failure,
-  }) =>
-      throw _privateConstructorUsedError;
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? loading,
-    TResult Function()? success,
-    TResult Function(String message)? failure,
-    required TResult orElse(),
-  }) =>
-      throw _privateConstructorUsedError;
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(_Initial value) initial,
-    required TResult Function(CreateLoading value) loading,
-    required TResult Function(CreateSuccess value) success,
-    required TResult Function(CreateFailure value) failure,
-  }) =>
-      throw _privateConstructorUsedError;
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(_Initial value)? initial,
-    TResult? Function(CreateLoading value)? loading,
-    TResult? Function(CreateSuccess value)? success,
-    TResult? Function(CreateFailure value)? failure,
-  }) =>
-      throw _privateConstructorUsedError;
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(_Initial value)? initial,
-    TResult Function(CreateLoading value)? loading,
-    TResult Function(CreateSuccess value)? success,
-    TResult Function(CreateFailure value)? failure,
-    required TResult orElse(),
-  }) =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $CreateuserStateCopyWith<$Res> {
-  factory $CreateuserStateCopyWith(
-          CreateuserState value, $Res Function(CreateuserState) then) =
-      _$CreateuserStateCopyWithImpl<$Res, CreateuserState>;
-}
-
-/// @nodoc
-class _$CreateuserStateCopyWithImpl<$Res, $Val extends CreateuserState>
-    implements $CreateuserStateCopyWith<$Res> {
-  _$CreateuserStateCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-}
-
-/// @nodoc
-abstract class _$$InitialImplCopyWith<$Res> {
-  factory _$$InitialImplCopyWith(
-          _$InitialImpl value, $Res Function(_$InitialImpl) then) =
-      __$$InitialImplCopyWithImpl<$Res>;
-}
-
-/// @nodoc
-class __$$InitialImplCopyWithImpl<$Res>
-    extends _$CreateuserStateCopyWithImpl<$Res, _$InitialImpl>
-    implements _$$InitialImplCopyWith<$Res> {
-  __$$InitialImplCopyWithImpl(
-      _$InitialImpl _value, $Res Function(_$InitialImpl) _then)
-      : super(_value, _then);
-}
-
-/// @nodoc
-
-class _$InitialImpl implements _Initial {
-  const _$InitialImpl();
-
-  @override
-  String toString() {
-    return 'CreateuserState.initial()';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$InitialImpl);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() loading,
-    required TResult Function() success,
-    required TResult Function(String message) failure,
-  }) {
-    return initial();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? loading,
-    TResult? Function()? success,
-    TResult? Function(String message)? failure,
-  }) {
-    return initial?.call();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? loading,
-    TResult Function()? success,
-    TResult Function(String message)? failure,
-    required TResult orElse(),
-  }) {
-    if (initial != null) {
-      return initial();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(_Initial value) initial,
-    required TResult Function(CreateLoading value) loading,
-    required TResult Function(CreateSuccess value) success,
-    required TResult Function(CreateFailure value) failure,
-  }) {
-    return initial(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(_Initial value)? initial,
-    TResult? Function(CreateLoading value)? loading,
-    TResult? Function(CreateSuccess value)? success,
-    TResult? Function(CreateFailure value)? failure,
-  }) {
-    return initial?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(_Initial value)? initial,
-    TResult Function(CreateLoading value)? loading,
-    TResult Function(CreateSuccess value)? success,
-    TResult Function(CreateFailure value)? failure,
-    required TResult orElse(),
-  }) {
-    if (initial != null) {
-      return initial(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class _Initial implements CreateuserState {
-  const factory _Initial() = _$InitialImpl;
-}
-
-/// @nodoc
-abstract class _$$CreateLoadingImplCopyWith<$Res> {
-  factory _$$CreateLoadingImplCopyWith(
-          _$CreateLoadingImpl value, $Res Function(_$CreateLoadingImpl) then) =
-      __$$CreateLoadingImplCopyWithImpl<$Res>;
-}
-
-/// @nodoc
-class __$$CreateLoadingImplCopyWithImpl<$Res>
-    extends _$CreateuserStateCopyWithImpl<$Res, _$CreateLoadingImpl>
-    implements _$$CreateLoadingImplCopyWith<$Res> {
-  __$$CreateLoadingImplCopyWithImpl(
-      _$CreateLoadingImpl _value, $Res Function(_$CreateLoadingImpl) _then)
-      : super(_value, _then);
-}
-
-/// @nodoc
-
-class _$CreateLoadingImpl implements CreateLoading {
-  const _$CreateLoadingImpl();
-
-  @override
-  String toString() {
-    return 'CreateuserState.loading()';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$CreateLoadingImpl);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() loading,
-    required TResult Function() success,
-    required TResult Function(String message) failure,
-  }) {
-    return loading();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? loading,
-    TResult? Function()? success,
-    TResult? Function(String message)? failure,
-  }) {
-    return loading?.call();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? loading,
-    TResult Function()? success,
-    TResult Function(String message)? failure,
-    required TResult orElse(),
-  }) {
-    if (loading != null) {
-      return loading();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(_Initial value) initial,
-    required TResult Function(CreateLoading value) loading,
-    required TResult Function(CreateSuccess value) success,
-    required TResult Function(CreateFailure value) failure,
-  }) {
-    return loading(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(_Initial value)? initial,
-    TResult? Function(CreateLoading value)? loading,
-    TResult? Function(CreateSuccess value)? success,
-    TResult? Function(CreateFailure value)? failure,
-  }) {
-    return loading?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(_Initial value)? initial,
-    TResult Function(CreateLoading value)? loading,
-    TResult Function(CreateSuccess value)? success,
-    TResult Function(CreateFailure value)? failure,
-    required TResult orElse(),
-  }) {
-    if (loading != null) {
-      return loading(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class CreateLoading implements CreateuserState {
-  const factory CreateLoading() = _$CreateLoadingImpl;
-}
-
-/// @nodoc
-abstract class _$$CreateSuccessImplCopyWith<$Res> {
-  factory _$$CreateSuccessImplCopyWith(
-          _$CreateSuccessImpl value, $Res Function(_$CreateSuccessImpl) then) =
-      __$$CreateSuccessImplCopyWithImpl<$Res>;
-}
-
-/// @nodoc
-class __$$CreateSuccessImplCopyWithImpl<$Res>
-    extends _$CreateuserStateCopyWithImpl<$Res, _$CreateSuccessImpl>
-    implements _$$CreateSuccessImplCopyWith<$Res> {
-  __$$CreateSuccessImplCopyWithImpl(
-      _$CreateSuccessImpl _value, $Res Function(_$CreateSuccessImpl) _then)
-      : super(_value, _then);
-}
-
-/// @nodoc
-
-class _$CreateSuccessImpl implements CreateSuccess {
-  const _$CreateSuccessImpl();
-
-  @override
-  String toString() {
-    return 'CreateuserState.success()';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$CreateSuccessImpl);
-  }
-
-  @override
-  int get hashCode => runtimeType.hashCode;
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() loading,
-    required TResult Function() success,
-    required TResult Function(String message) failure,
-  }) {
-    return success();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? loading,
-    TResult? Function()? success,
-    TResult? Function(String message)? failure,
-  }) {
-    return success?.call();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? loading,
-    TResult Function()? success,
-    TResult Function(String message)? failure,
-    required TResult orElse(),
-  }) {
-    if (success != null) {
-      return success();
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(_Initial value) initial,
-    required TResult Function(CreateLoading value) loading,
-    required TResult Function(CreateSuccess value) success,
-    required TResult Function(CreateFailure value) failure,
-  }) {
-    return success(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(_Initial value)? initial,
-    TResult? Function(CreateLoading value)? loading,
-    TResult? Function(CreateSuccess value)? success,
-    TResult? Function(CreateFailure value)? failure,
-  }) {
-    return success?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(_Initial value)? initial,
-    TResult Function(CreateLoading value)? loading,
-    TResult Function(CreateSuccess value)? success,
-    TResult Function(CreateFailure value)? failure,
-    required TResult orElse(),
-  }) {
-    if (success != null) {
-      return success(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class CreateSuccess implements CreateuserState {
-  const factory CreateSuccess() = _$CreateSuccessImpl;
-}
-
-/// @nodoc
-abstract class _$$CreateFailureImplCopyWith<$Res> {
-  factory _$$CreateFailureImplCopyWith(
-          _$CreateFailureImpl value, $Res Function(_$CreateFailureImpl) then) =
-      __$$CreateFailureImplCopyWithImpl<$Res>;
-  @useResult
-  $Res call({String message});
-}
-
-/// @nodoc
-class __$$CreateFailureImplCopyWithImpl<$Res>
-    extends _$CreateuserStateCopyWithImpl<$Res, _$CreateFailureImpl>
-    implements _$$CreateFailureImplCopyWith<$Res> {
-  __$$CreateFailureImplCopyWithImpl(
-      _$CreateFailureImpl _value, $Res Function(_$CreateFailureImpl) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? message = null,
-  }) {
-    return _then(_$CreateFailureImpl(
-      message: null == message
-          ? _value.message
-          : message // ignore: cast_nullable_to_non_nullable
-              as String,
-    ));
-  }
-}
-
-/// @nodoc
-
-class _$CreateFailureImpl implements CreateFailure {
-  const _$CreateFailureImpl({required this.message});
-
-  @override
-  final String message;
-
-  @override
-  String toString() {
-    return 'CreateuserState.failure(message: $message)';
-  }
-
-  @override
-  bool operator ==(Object other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$CreateFailureImpl &&
-            (identical(other.message, message) || other.message == message));
-  }
-
-  @override
-  int get hashCode => Object.hash(runtimeType, message);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$CreateFailureImplCopyWith<_$CreateFailureImpl> get copyWith =>
-      __$$CreateFailureImplCopyWithImpl<_$CreateFailureImpl>(this, _$identity);
-
-  @override
-  @optionalTypeArgs
-  TResult when<TResult extends Object?>({
-    required TResult Function() initial,
-    required TResult Function() loading,
-    required TResult Function() success,
-    required TResult Function(String message) failure,
-  }) {
-    return failure(message);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? whenOrNull<TResult extends Object?>({
-    TResult? Function()? initial,
-    TResult? Function()? loading,
-    TResult? Function()? success,
-    TResult? Function(String message)? failure,
-  }) {
-    return failure?.call(message);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeWhen<TResult extends Object?>({
-    TResult Function()? initial,
-    TResult Function()? loading,
-    TResult Function()? success,
-    TResult Function(String message)? failure,
-    required TResult orElse(),
-  }) {
-    if (failure != null) {
-      return failure(message);
-    }
-    return orElse();
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult map<TResult extends Object?>({
-    required TResult Function(_Initial value) initial,
-    required TResult Function(CreateLoading value) loading,
-    required TResult Function(CreateSuccess value) success,
-    required TResult Function(CreateFailure value) failure,
-  }) {
-    return failure(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult? mapOrNull<TResult extends Object?>({
-    TResult? Function(_Initial value)? initial,
-    TResult? Function(CreateLoading value)? loading,
-    TResult? Function(CreateSuccess value)? success,
-    TResult? Function(CreateFailure value)? failure,
-  }) {
-    return failure?.call(this);
-  }
-
-  @override
-  @optionalTypeArgs
-  TResult maybeMap<TResult extends Object?>({
-    TResult Function(_Initial value)? initial,
-    TResult Function(CreateLoading value)? loading,
-    TResult Function(CreateSuccess value)? success,
-    TResult Function(CreateFailure value)? failure,
-    required TResult orElse(),
-  }) {
-    if (failure != null) {
-      return failure(this);
-    }
-    return orElse();
-  }
-}
-
-abstract class CreateFailure implements CreateuserState {
-  const factory CreateFailure({required final String message}) =
-      _$CreateFailureImpl;
-
-  String get message;
-  @JsonKey(ignore: true)
-  _$$CreateFailureImplCopyWith<_$CreateFailureImpl> get copyWith =>
-      throw _privateConstructorUsedError;
-}
