--- lib/features/add_user_feature/ui/widgets/create_user_text.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:rikaz_team/core/theming/styles.dart';
-
-class CreateUserText extends StatelessWidget {
-  const CreateUserText({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-          "Create User",
-          style: TextStyles.font32BlueBold,
-        );
-  }
-}
