+++ lib/features/login_feature/data/apis/api_service.dart
@@ -1,6 +1,5 @@
 import 'package:dio/dio.dart';
 import 'package:rikaz_team/core/network/api_constances.dart';
-import 'package:rikaz_team/features/add_user_feature/data/models/create_user_request.dart';
 import 'package:rikaz_team/features/login_feature/data/models/login_request.dart';
 
 class ApiService {
@@ -16,12 +15,12 @@ class ApiService {
     return response.data;
   }
 
-  Future<Map<String , dynamic>> createuser(CreateUserRequest createUserRequest) async {
-    var response = await _dio.post(
-      ApiConstances.createUserUrl,
-      data: createUserRequest.toJson(),
-    );
-    return response.data;
-  }
+  // Future<Map<String , dynamic>> createuser(CreateUserRequest createUserRequest) async {
+  //   var response = await _dio.post(
+  //     ApiConstances.createUserUrl,
+  //     data: createUserRequest.toJson(),
+  //   );
+  //   return response.data;
+  // }
 
 }
