+++ lib/features/login_feature/ui/widgets/dont_have_account.dart
@@ -1,11 +1,6 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:rikaz_team/core/services/services_locator.dart';
 import 'package:rikaz_team/core/theming/styles.dart';
-import 'package:rikaz_team/features/add_user_feature/ui/add_user_screen.dart';
-
-import '../../../add_user_feature/logic/bloc/createuser_bloc.dart';
 
 class DontHaveAccountText extends StatelessWidget {
   const DontHaveAccountText({super.key});
@@ -27,13 +22,13 @@ class DontHaveAccountText extends StatelessWidget {
               style: TextStyles.font13BlueRegular.copyWith(fontSize: 16),
               recognizer: TapGestureRecognizer()
                 ..onTap = () {
-                  Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                          builder: (context) => BlocProvider(
-                                create: (context) => sl<CreateuserBloc>(),
-                                child: const AddUserScreen(),
-                              )));
+                  // Navigator.push(
+                  //     context,
+                  //     MaterialPageRoute(
+                  //         builder: (context) => BlocProvider(
+                  //               create: (context) => sl<CreateuserBloc>(),
+                  //               child: const AddUserScreen(),
+                  //             )));
                 },
             ),
           ],
