+++ lib/features/login_feature/ui/widgets/home_screen.dart
@@ -5,8 +5,9 @@ class HomeScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body:  Center(
+    return Scaffold(
+      appBar: AppBar(),
+      body: const Center(
         child: Text("Home Screeen !!!!"),
       ),
     );
